requirements.txt
pandas==2.0.3
flask==2.2.2
requests==2.31.0
numpy==1.24.3
Werkzeug==2.2.3
google-auth==2.35.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.26.0
db-dtypes==1.3.0